add connect button to simplebox example
