--- examples/lobby/Cargo.toml
@@ -1,40 +1,0 @@
-[package]
-name = "lobby"
-version = "0.1.0"
-authors = ["Charles Bournhonesque <charlesbour@gmail.com>"]
-edition = "2021"
-rust-version = "1.65"
-description = "Examples for the lightyearServer-client networking library for the Bevy game engine"
-readme = "README.md"
-repository = "https://github.com/cBournhonesque/lightyear"
-keywords = ["bevy", "multiplayer", "networking", "netcode", "gamedev"]
-categories = ["game-development", "network-programming"]
-license = "MIT OR Apache-2.0"
-publish = false
-
-[features]
-metrics = ["lightyear/metrics", "dep:metrics-exporter-prometheus"]
-mock_time = ["lightyear/mock_time"]
-
-[dependencies]
-lightyear = { path = "../../lightyear", features = [
-    "steam",
-    "webtransport",
-    "websocket",
-    "render",
-] }
-async-compat = "0.2.3"
-serde = { version = "1.0.188", features = ["derive"] }
-anyhow = { version = "1.0.75", features = [] }
-tracing = "0.1"
-tracing-subscriber = "0.3.17"
-bevy = { version = "0.13", features = ["bevy_core_pipeline"] }
-bevy_mod_picking = { version = "0.18.2", features = ["backend_bevy_ui"] }
-derive_more = { version = "0.99", features = ["add", "mul"] }
-rand = "0.8.1"
-clap = { version = "4.4", features = ["derive"] }
-mock_instant = "0.4"
-metrics-exporter-prometheus = { version = "0.13.0", optional = true }
-bevy-inspector-egui = "0.23"
-cfg-if = "1.0.0"
-crossbeam-channel = "0.5.11"
