--- examples/lobby/assets/settings.ron
@@ -1,57 +1,0 @@
-Settings(
-    client: ClientSettings(
-        inspector: true,
-        client_id: 0,
-        client_port: 0, // the OS will assign a random open port
-        server_addr: "127.0.0.1",
-        conditioner: Some(Conditioner(
-            latency_ms: 200,
-            jitter_ms: 20,
-            packet_loss: 0.05
-        )),
-        server_port: 5000,
-        transport: WebTransport(
-            // this is only needed for wasm, the self-signed certificates are only valid for 2 weeks
-            // the server will print the certificate digest on startup
-            certificate_digest: "a71e935276ac0e37db88bfa702b3ece66f30b3ad4f6117c12b782fe0f2817f9f",
-        ),
-        // server_port: 5001,
-        // transport: Udp,
-        // server_port: 5002,
-        // transport: WebSocket,
-        // server_port: 5003,
-        // transport: Steam(
-        //     app_id: 480,
-        // )
-    ),
-    server: ServerSettings(
-        headless: true,
-        inspector: false,
-        conditioner: Some(Conditioner(
-            latency_ms: 200,
-            jitter_ms: 20,
-            packet_loss: 0.05
-        )),
-        transport: [
-            WebTransport(
-                local_port: 5000
-            ),
-            Udp(
-                local_port: 5001
-            ),
-            WebSocket(
-                local_port: 5002
-            ),
-            // Steam(
-            //     app_id: 480,
-            //     server_ip: "0.0.0.0",
-            //     game_port: 5003,
-            //     query_port: 27016,
-            // ),
-        ],
-    ),
-    shared: SharedSettings(
-        protocol_id: 0,
-        private_key: (0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0),
-    )
-)
