--- examples/lobby/index.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
-    <title>Bevy game</title>
-</head>
-</html>
