--- examples/lobby/src/protocol.rs
@@ -1,142 +1,0 @@
-//! This file contains the shared [`Protocol`] that defines the messages that can be sent between the client and server.
-//!
-//! You will need to define the [`Components`], [`Messages`] and [`Inputs`] that make up the protocol.
-//! You can use the `#[protocol]` attribute to specify additional behaviour:
-//! - how entities contained in the message should be mapped from the remote world to the local world
-//! - how the component should be synchronized between the `Confirmed` entity and the `Predicted`/`Interpolated` entity
-use std::ops::Mul;
-
-use bevy::ecs::entity::MapEntities;
-use bevy::prelude::Parent;
-use bevy::prelude::{
-    default, Bundle, Color, Component, Deref, DerefMut, Entity, EntityMapper, Vec2,
-};
-use derive_more::{Add, Mul};
-use serde::{Deserialize, Serialize};
-
-use lightyear::prelude::*;
-
-// Player
-#[derive(Bundle)]
-pub(crate) struct PlayerBundle {
-    id: PlayerId,
-    position: PlayerPosition,
-    color: PlayerColor,
-}
-
-impl PlayerBundle {
-    pub(crate) fn new(id: ClientId, position: Vec2) -> Self {
-        // Generate pseudo random color from client id.
-        let h = (((id.to_bits().wrapping_mul(30)) % 360) as f32) / 360.0;
-        let s = 0.8;
-        let l = 0.5;
-        let color = Color::hsl(h, s, l);
-        Self {
-            id: PlayerId(id),
-            position: PlayerPosition(position),
-            color: PlayerColor(color),
-        }
-    }
-}
-
-// Components
-
-#[derive(Component, Serialize, Deserialize, Clone, Debug, PartialEq)]
-pub struct PlayerId(ClientId);
-
-#[derive(Component, Serialize, Deserialize, Clone, Debug, PartialEq, Deref, DerefMut, Add, Mul)]
-pub struct PlayerPosition(Vec2);
-
-impl Mul<f32> for &PlayerPosition {
-    type Output = PlayerPosition;
-
-    fn mul(self, rhs: f32) -> Self::Output {
-        PlayerPosition(self.0 * rhs)
-    }
-}
-
-#[derive(Component, Deserialize, Serialize, Clone, Debug, PartialEq)]
-pub struct PlayerColor(pub(crate) Color);
-
-// Example of a component that contains an entity.
-// This component, when replicated, needs to have the inner entity mapped from the Server world
-// to the client World.
-// This can be done by adding a `#[message(custom_map)]` attribute to the component, and then
-// deriving the `MapEntities` trait for the component.
-#[derive(Component, Deserialize, Serialize, Clone, Debug, PartialEq)]
-pub struct PlayerParent(Entity);
-
-impl MapEntities for PlayerParent {
-    fn map_entities<M: EntityMapper>(&mut self, entity_mapper: &mut M) {
-        self.0 = entity_mapper.map_entity(self.0);
-    }
-}
-
-#[component_protocol(protocol = "MyProtocol")]
-pub enum Components {
-    #[protocol(sync(mode = "once"))]
-    PlayerId(PlayerId),
-    #[protocol(sync(mode = "full"))]
-    PlayerPosition(PlayerPosition),
-    #[protocol(sync(mode = "once"))]
-    PlayerColor(PlayerColor),
-}
-
-// Channels
-
-#[derive(Channel)]
-pub struct Channel1;
-
-// Messages
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
-pub struct Message1(pub usize);
-
-#[message_protocol(protocol = "MyProtocol")]
-pub enum Messages {
-    Message1(Message1),
-}
-
-// Inputs
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone)]
-pub struct Direction {
-    pub(crate) up: bool,
-    pub(crate) down: bool,
-    pub(crate) left: bool,
-    pub(crate) right: bool,
-}
-
-impl Direction {
-    pub(crate) fn is_none(&self) -> bool {
-        !self.up && !self.down && !self.left && !self.right
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
-pub enum Inputs {
-    Direction(Direction),
-    Delete,
-    Spawn,
-    None,
-}
-
-impl UserAction for Inputs {}
-
-// Protocol
-
-protocolize! {
-    Self = MyProtocol,
-    Message = Messages,
-    Component = Components,
-    Input = Inputs,
-}
-
-pub(crate) fn protocol() -> MyProtocol {
-    let mut protocol = MyProtocol::default();
-    protocol.add_channel::<Channel1>(ChannelSettings {
-        mode: ChannelMode::OrderedReliable(ReliableSettings::default()),
-        ..default()
-    });
-    protocol
-}
