+++ examples/simple_box/Cargo.toml
@@ -18,10 +18,10 @@ mock_time = ["lightyear/mock_time"]
 
 [dependencies]
 lightyear = { path = "../../lightyear", features = [
-  "steam",
-  "webtransport",
-  "websocket",
-  "render",
+    "steam",
+    "webtransport",
+    "websocket",
+    "render",
 ] }
 async-compat = "0.2.3"
 serde = { version = "1.0.188", features = ["derive"] }
@@ -29,6 +29,7 @@ anyhow = { version = "1.0.75", features = [] }
 tracing = "0.1"
 tracing-subscriber = "0.3.17"
 bevy = { version = "0.13", features = ["bevy_core_pipeline"] }
+bevy_mod_picking = { version = "0.18.2", features = ["backend_bevy_ui"] }
 derive_more = { version = "0.99", features = ["add", "mul"] }
 rand = "0.8.1"
 clap = { version = "4.4", features = ["derive"] }
