+++ examples/simple_box/src/client.rs
@@ -11,6 +11,8 @@ use bevy::app::PluginGroupBuilder;
 use bevy::prelude::*;
 use bevy::time::common_conditions::on_timer;
 use bevy::utils::Duration;
+use bevy_mod_picking::picking_core::Pickable;
+use bevy_mod_picking::prelude::{Click, On, Pointer};
 
 pub use lightyear::prelude::client::*;
 use lightyear::prelude::*;
@@ -24,7 +26,7 @@ pub struct ExampleClientPlugin;
 
 impl Plugin for ExampleClientPlugin {
     fn build(&self, app: &mut App) {
-        app.add_systems(Startup, init);
+        app.add_systems(Startup, spawn_connect_button);
         app.add_systems(PreUpdate, handle_connection.after(MainSet::Receive));
         // Inputs have to be buffered in the FixedPreUpdate schedule
         app.add_systems(
@@ -40,15 +42,16 @@ impl Plugin for ExampleClientPlugin {
                 receive_entity_despawn,
                 handle_predicted_spawn,
                 handle_interpolated_spawn,
+                button_system,
             ),
         );
+        app.add_systems(OnEnter(NetworkingState::Disconnected), on_disconnect);
     }
 }
 
-/// Startup system for the client.
-pub(crate) fn init(mut client: ClientConnectionParam) {
-    let _ = client.connect();
-}
+/// Component to identify the text displaying the client id
+#[derive(Component)]
+pub struct ClientIdText;
 
 /// Listen for events to know when the client is connected, and spawn a text entity
 /// to display the client id
@@ -58,13 +61,16 @@ pub(crate) fn handle_connection(
 ) {
     for event in connection_event.read() {
         let client_id = event.client_id();
-        commands.spawn(TextBundle::from_section(
-            format!("Client {}", client_id),
-            TextStyle {
-                font_size: 30.0,
-                color: Color::WHITE,
-                ..default()
-            },
+        commands.spawn((
+            TextBundle::from_section(
+                format!("Client {}", client_id),
+                TextStyle {
+                    font_size: 30.0,
+                    color: Color::WHITE,
+                    ..default()
+                },
+            ),
+            ClientIdText,
         ));
     }
 }
@@ -166,5 +172,109 @@ pub(crate) fn handle_interpolated_spawn(
 ) {
     for mut color in interpolated.iter_mut() {
         color.0.set_s(0.1);
+    }
+}
+
+/// Create a button that allow you to connect/disconnect to a server
+pub(crate) fn spawn_connect_button(mut commands: Commands) {
+    commands
+        .spawn((
+            NodeBundle {
+                style: Style {
+                    width: Val::Percent(100.0),
+                    height: Val::Percent(100.0),
+                    align_items: AlignItems::FlexEnd,
+                    justify_content: JustifyContent::FlexEnd,
+                    flex_direction: FlexDirection::Row,
+                    ..default()
+                },
+                ..default()
+            },
+            Pickable::IGNORE,
+        ))
+        .with_children(|parent| {
+            parent
+                .spawn((
+                    ButtonBundle {
+                        style: Style {
+                            width: Val::Px(150.0),
+                            height: Val::Px(65.0),
+                            border: UiRect::all(Val::Px(5.0)),
+                            // horizontally center child text
+                            justify_content: JustifyContent::Center,
+                            // vertically center child text
+                            align_items: AlignItems::Center,
+                            ..default()
+                        },
+                        border_color: BorderColor(Color::BLACK),
+                        background_color: Color::rgb(0.15, 0.15, 0.15).into(),
+                        ..default()
+                    },
+                    On::<Pointer<Click>>::run(|| {}),
+                ))
+                .with_children(|parent| {
+                    parent.spawn((
+                        TextBundle::from_section(
+                            "Connect",
+                            TextStyle {
+                                font_size: 20.0,
+                                color: Color::rgb(0.9, 0.9, 0.9),
+                                ..default()
+                            },
+                        ),
+                        Pickable::IGNORE,
+                    ));
+                });
+        });
+}
+
+/// Remove all entities when the client disconnect
+fn on_disconnect(
+    mut commands: Commands,
+    player_entities: Query<Entity, With<PlayerId>>,
+    debug_text: Query<Entity, With<ClientIdText>>,
+) {
+    for entity in player_entities.iter() {
+        commands.entity(entity).despawn_recursive();
+    }
+    for entity in debug_text.iter() {
+        commands.entity(entity).despawn_recursive();
+    }
+}
+
+///  System that will assign a callback to the 'Connect' button depending on the connection state.
+fn button_system(
+    mut interaction_query: Query<(Entity, &Children, &mut On<Pointer<Click>>), With<Button>>,
+    mut text_query: Query<&mut Text>,
+    state: Res<State<NetworkingState>>,
+) {
+    if state.is_changed() {
+        for (entity, children, mut on_click) in &mut interaction_query {
+            let mut text = text_query.get_mut(children[0]).unwrap();
+            match state.get() {
+                NetworkingState::Disconnected => {
+                    text.sections[0].value = "Connect".to_string();
+                    *on_click =
+                        On::<Pointer<Click>>::run(|mut connection: ClientConnectionParam| {
+                            let _ = connection
+                                .connect()
+                                .inspect_err(|e| error!("Failed to connect: {e:?}"));
+                        });
+                }
+                NetworkingState::Connecting => {
+                    text.sections[0].value = "Connecting".to_string();
+                    *on_click = On::<Pointer<Click>>::run(|| {});
+                }
+                NetworkingState::Connected => {
+                    text.sections[0].value = "Disconnect".to_string();
+                    *on_click =
+                        On::<Pointer<Click>>::run(|mut connection: ClientConnectionParam| {
+                            let _ = connection
+                                .disconnect()
+                                .inspect_err(|e| error!("Failed to disconnect: {e:?}"));
+                        });
+                }
+            };
+        }
     }
 }
