+++ examples/simple_box/src/shared.rs
@@ -6,6 +6,7 @@
 use bevy::prelude::*;
 use bevy::render::RenderPlugin;
 use bevy::utils::Duration;
+use bevy_mod_picking::DefaultPickingPlugins;
 
 use lightyear::prelude::client::*;
 use lightyear::prelude::*;
@@ -30,6 +31,7 @@ pub struct SharedPlugin;
 impl Plugin for SharedPlugin {
     fn build(&self, app: &mut App) {
         if app.is_plugin_added::<RenderPlugin>() {
+            app.add_plugins(DefaultPickingPlugins);
             app.add_systems(Startup, init);
             app.add_systems(Update, draw_boxes);
         }
